[settings] Moved the settings files to a subfolder to avoid conflict with `pip install --editable` and other tools
