+++ .travis.yml
@@ -47,7 +47,6 @@ matrix:
       - DATABASE=_postgres
 
 before_script:
-  - export DJANGO_SETTINGS_MODULE=settings
   - export PYTHONPATH=$PYTHONPATH:$(pwd)
   # check gis
   - gdal-config --version
@@ -70,7 +69,7 @@ install:
   - pip install Django~=${DJANGO}.0
 
 script:
-  - pytest --create-db --reuse-db --no-migrations --ds=settings${DATABASE}
+  - pytest --create-db --reuse-db --no-migrations --ds=ddf_settings.settings${DATABASE}
 
 after_success:
   - coveralls
