+++ manage.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 try:
-    import settings # Assumed to be in the same directory.
+    import ddf_settings.settings # Assumed to be in the same directory.
 except ImportError:
     import sys
     sys.stderr.write("Error: Can't find the file 'settings.py' in the directory containing %r. It appears you've customized things.\nYou'll have to run django-admin.py, passing it your settings module.\n(If the file settings.py does indeed exist, it's causing an ImportError somehow.)\n" % __file__)
@@ -8,6 +8,6 @@ except ImportError:
 
 if __name__ == "__main__":
     import os
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "ddf_settings.settings")
     from django.core.management import execute_from_command_line
     execute_from_command_line()
