+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
-DJANGO_SETTINGS_MODULE = settings
+DJANGO_SETTINGS_MODULE = ddf_settings.settings
 addopts = --pyargs --create-db --reuse-db --no-migrations
 python_files = test_*.py
